misc cleanup and reorg
- reorganize types
- remove some refs to old name
